feat(): connect GradeView to db

+++ Academix/Models/Thamso.cs
@@ -7,5 +7,5 @@ public partial class Thamso
 {
     public string Tenthamso { get; set; } = null!;
 
-    public float? Giatri { get; set; }
+    public double? Giatri { get; set; }
 }
