+++ Academix/ViewModels/SemesterReportItemViewModel.cs
@@ -39,8 +39,8 @@ namespace Academix.ViewModels
             }
         }
 
-        private float _passingRate;
-        public float PassingRate
+        private double _passingRate;
+        public double PassingRate
         {
             get => _passingRate;
             set
@@ -50,7 +50,7 @@ namespace Academix.ViewModels
             }
         }
 
-        public SemesterReportItemViewModel(Lop cl, int count, float passingRate)
+        public SemesterReportItemViewModel(Lop cl, int count, double passingRate)
         {
             _class = cl;
             _count = count;
